Now compiling and running on windows, but something is very wrong with ComponentModel.

+++ engine/include/rev/NurbsCurve.h
@@ -7,6 +7,7 @@
 #include <gsl/gsl_assert>
 #include <gsl/span>
 #include <limits>
+#include <vector>
 
 namespace rev {
 
@@ -29,7 +30,7 @@ public:
         , _controlPoints(controlPoints.begin(), controlPoints.end())
     {
         Expects(order > 1);
-        Expects(controlPoints.size() >= order);
+        Expects(static_cast<size_t>(controlPoints.size()) >= order);
         Expects(knots.size() == (controlPoints.size() + order));
     }
